requirements.txt
numpy
selenium==4.7.2
webdriver_manager==3.8.5
matplotlib
missingno==0.5.1
pandas
python-dotenv